System test additions
 - run --userns=keep-id: confirm that $HOME gets set (#8013)

 - inspect: confirm that JSON output is a sane number of
   lines (10 or more), not an unreadable one-liner (#8011
   and #8021). Do so with image, pod, network, volume
   because the code paths might be different.

 - cgroups: confirm that 'run' preserves cgroup manager (#7970)

 - sdnotify: reenable tests, and hope CI doesn't hang. This
   test was disabled on August 18 because CI jobs were hanging
   and timing out. My suspicion was that it was #7316, which
   in turn seems to have hinged on conmon #182. The latter
   was merged on Sep 16, so let's cross our fingers and see
   what happens.

Also: remove inaccurate warning from a networking test.

And, wow, fix is_cgroupsv2(), it has never actually worked.

+++ test/system/500-networking.bats
@@ -90,7 +90,12 @@ load helpers
     run_podman network create --subnet "${mysubnet}.0/24" $mynetname
     is "$output" ".*/cni/net.d/$mynetname.conflist" "output of 'network create'"
 
-    # WARNING: this pulls a ~100MB image from quay.io, hence is slow/flaky
+    # (Assert that output is formatted, not a one-line blob: #8011)
+    run_podman network inspect $mynetname
+    if [[ "${#lines[*]}" -lt 5 ]]; then
+        die "Output from 'pod inspect' is only ${#lines[*]} lines; see #8011"
+    fi
+
     run_podman run --rm --network $mynetname $IMAGE ip a
     is "$output" ".* inet ${mysubnet}\.2/24 brd ${mysubnet}\.255 " \
        "sdfsdf"
